Adds a supasquad program
