+++ web/docs/handbook/contributing.mdx
@@ -4,42 +4,22 @@ title: Contributing
 description: Want to help?
 ---
 
+import Sponsors from '../../src/components/Sponsors'
+
 ## How to contribute
 
 Want to contribute? Why not jump into our GitHub repo and:
 
+- [Sponsor Supabase](https://github.com/sponsors/supabase).
+- [Answer Discussions](https://github.com/supabase/supabase/discussions).
 - Submit an issue.
 - Report a performance issue or a part of the documentation that you find confusing.
 - Create a pull request.
-- Try one of our products and give feedback.
-- Translate this website.
-- [Vote on a new client library](https://github.com/supabase/supabase/issues/5) for your favorite language.
+- [Translate our Readme](https://github.com/supabase/supabase/issues/1341).
+- Try our products and give feedback.
 - Spread the word if you like what we are doing.
-- Add your project or company to our showcase.
 
 
-## Contributors
+## Sponsors
 
-To add yourself as a contributor:
-
-Open [this page](https://github.com/supabase/supabase/blob/master/web/docs/handbook/contributing.mdx) on GitHub and then click the pencil icon.
-
-![Edit this page](/img/edit-contributors.png)
-
-Enter your GitHub profile name and link on a new line.
-
-![Add your Github handle](/img/add-contributor.png)
-
-Enter a title for your pull request and click "Propose Changes":
-
-![Submit PR](/img/propose-changes.png)
-
-One of the Supabase maintainers will approve your PR in the next few hours.
-
-<!-- Add your github handle below this line. Every name should be on a new line -->
-[awalias](https://github.com/awalias).
-[kiwicopple](https://github.com/kiwicopple).
-[abhinav-TB](https://github.com/abhinav-TB).
-[jacobfenger](https://github.com/jacobfenger).
-<!-- Add your github above this line. Every name should be on a new line -->
-
+<Sponsors />
