+++ web/docs/handbook/supasquad.mdx
@@ -1,0 +1,76 @@
+---
+id: supasquad
+title: SupaSquad
+description: Supabase Advocate program
+---
+
+The SupaSquad is an official Supabase advocate program where community members help build and manage the Supabase community.
+
+- Official recognition in the Supabase community.
+- Direct connection to the Supabase team.
+- Limited members who help steer the Supabase community.
+
+## Requirements
+
+As a member of the Squad, you can choose the approach you feel you'll be able to provide the most value with.
+You can help in one of five ways:
+
+### Maintainer
+
+Help maintain Supabase repositories. This includes building client librararies, managing issues, and fixing bugs.
+
+### Expert
+
+Answer user questions on Github [Discussions](https://github.com/supabase/supabase/discussions), Discord, and on various other social platforms.
+
+### Advocate
+
+Spread the word on social channels and help to answer Supabase-related questions in the broader community and social channels
+
+### Builder
+
+Build Supabase examples, blog about them, and them add to the [SupaSquad GitHub Org](https://github.com/supasquad).
+
+### Author
+
+Write guest blog posts, create documentation, and help Supabase global expansion through translation.
+
+## Benefits for SupaSquad members
+
+- Access to a Supabase Discord channel providing direct communication with the team, Discord badges, and elevated privileges.
+- Special AMA sessions with members of the Supabase team.
+- Monthly DevRel committee call with industry-leading Developer Advocates (many of whom are [angel investors](https://supabase.io/blog/2021/03/25/angels-of-supabase)), where you can learn from the best.
+- We'll help you build your audience by promoting content via the Supabase social channels.
+- Featured profile on Supabase website.
+- Early access to new features (and the opportunity to provide feedback to the team!).
+- Free credits that you can use for Squad efforts.
+- Direct access to members of the Supabase team for questions, suggestions, etc.
+- Help shape the future of the program.
+- Invited to the [SupaSquad GitHub Org](https://github.com/supasquad) to collaborate with the other squad members.
+- Exclusive Supabase Team swag drops usually exclusively reserved to the Supabase core team
+
+## How to join
+
+
+1. Apply to join the program using [this form](https://airtable.com/shr0FtLqLfhpuEya8).
+2. We will start taking interviewing members from `1 September 2021`.
+3. Entry is capped to 20 founding members so we can grow the community at a healthy pace and provide a great experience to all new users.
+4. Membership is valid for one year, and can be renewed every year.
+
+
+## FAQs
+
+<details>
+    <summary><i>Why are you only admitting 20 new members?</i></summary>
+    The entire Supabase team is only 20 people, so as you can imagine adding another 20 people sounds like a lot to us! We wish we could admit everyone who wanted to join. 
+    But we also want to make sure everyone who joins the Squad has an awesome experience. In the future we will probably expand the intake to include a monthly quota.
+</details>
+<details>
+    <summary><i>What is expected?</i></summary>
+    Mostly just enthusiasm. If you are interested in Open Source and want to get involved, the SupaSquad program is a great channel. You'll be given opportunities to contribute 
+    to the community in whatever ways match your skillset.
+</details>
+<details>
+    <summary><i>What if I become too busy to contribute?</i></summary>
+    No worries! The program isn't a job. It's just an opportunity to build your skillset and audience within the Supabase ecosystem.
+</details>
